[MCRAMC] Regenerated demo for mcramc__mcramc_ecc_testing.X
MH3-100314

+++ apps/mcramc/src/config/pic32cz_ca80_curiosity_ultra/peripheral/clock/plib_clock.c
@@ -52,7 +52,7 @@ static void PLL0_Initialize(void)
 {
     /* Enable Additional Voltage Regulator */
     SUPC_REGS->SUPC_VREGCTRL |= SUPC_VREGCTRL_AVREGEN(4U);
-    while ((SUPC_REGS->SUPC_STATUS & (4U << SUPC_STATUS_ADDVREGRDY_Pos)) != (4U << SUPC_STATUS_ADDVREGRDY_Pos))
+    while ((SUPC_REGS->SUPC_STATUS & ((uint32_t)4U << SUPC_STATUS_ADDVREGRDY_Pos)) != ((uint32_t)4U << SUPC_STATUS_ADDVREGRDY_Pos))
     {
         /* Do Nothing */
     }
